feat: improve mysql database searches by introducing json columns
