+++ node/src/main/java/eu/cloudnetservice/node/database/LocalDatabase.java
@@ -20,23 +20,14 @@ import eu.cloudnetservice.common.document.gson.JsonDocument;
 import eu.cloudnetservice.driver.database.Database;
 import java.util.Map;
 import java.util.function.BiConsumer;
-import java.util.function.BiPredicate;
 import lombok.NonNull;
 import org.jetbrains.annotations.Nullable;
 
 public interface LocalDatabase extends Database {
 
   /**
-   * Retrieves all entries that match the given filter predicate
-   *
-   * @param predicate the filter for the entries
-   * @return all entries that match the filter
-   */
-  @NonNull Map<String, JsonDocument> filter(@NonNull BiPredicate<String, JsonDocument> predicate);
-
-  /**
-   * Iterates over all entries in the database This option should not be used with big databases Use {@link
-   * #iterate(BiConsumer, int)}} instead
+   * Iterates over all entries in the database This option should not be used with big databases Use
+   * {@link #iterate(BiConsumer, int)}} instead
    *
    * @param consumer the consumer to pass the entries into
    */
