+++ node/src/main/java/eu/cloudnetservice/node/database/h2/H2Database.java
@@ -19,9 +19,16 @@ package eu.cloudnetservice.node.database.h2;
 import eu.cloudnetservice.common.document.gson.JsonDocument;
 import eu.cloudnetservice.node.database.sql.SQLDatabase;
 import eu.cloudnetservice.node.database.sql.SQLDatabaseProvider;
+import java.sql.ResultSet;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.WeakHashMap;
 import java.util.concurrent.ExecutorService;
+import java.util.function.BiConsumer;
 import lombok.NonNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -29,6 +36,204 @@ public final class H2Database extends SQLDatabase {
 
   public H2Database(@NonNull SQLDatabaseProvider provider, @NonNull String name, @NonNull ExecutorService executor) {
     super(provider, name, executor);
+
+    // create the table
+    provider.executeUpdate(String.format(
+      "CREATE TABLE IF NOT EXISTS `%s` (%s VARCHAR(64) PRIMARY KEY, %s TEXT);",
+      name,
+      TABLE_COLUMN_KEY,
+      TABLE_COLUMN_VAL));
+  }
+
+  @Override
+  public void close() {
+    this.databaseProvider.databaseCache().invalidate(this.name);
+  }
+
+  @Override
+  public boolean insert(@NonNull String key, @NonNull JsonDocument document) {
+    this.databaseProvider.databaseHandler().handleInsert(this, key, document);
+    return this.insertOrUpdate(key, document);
+  }
+
+  private boolean insert0(@NonNull String key, @NonNull JsonDocument document) {
+    return this.databaseProvider.executeUpdate(
+      "INSERT INTO `" + this.name + "` (" + TABLE_COLUMN_KEY + "," + TABLE_COLUMN_VAL + ") VALUES (?, ?);",
+      key, document.toString()
+    ) != -1;
+  }
+
+  public boolean update0(String key, JsonDocument document) {
+    return this.databaseProvider.executeUpdate(
+      "UPDATE `" + this.name + "` SET " + TABLE_COLUMN_VAL + "=? WHERE " + TABLE_COLUMN_KEY + "=?",
+      document.toString(), key
+    ) != -1;
+  }
+
+  public boolean insertOrUpdate(String key, JsonDocument document) {
+    return this.contains(key) ? this.update0(key, document) : this.insert0(key, document);
+  }
+
+  @Override
+  public boolean contains(@NonNull String key) {
+    return this.databaseProvider.executeQuery(
+      String.format("SELECT %s FROM `%s` WHERE %s = ?", TABLE_COLUMN_KEY, this.name, TABLE_COLUMN_KEY),
+      ResultSet::next,
+      false,
+      key);
+  }
+
+  @Override
+  public boolean delete(@NonNull String key) {
+    this.databaseProvider.databaseHandler().handleDelete(this, key);
+    return this.delete0(key);
+  }
+
+  public boolean delete0(String key) {
+    return this.databaseProvider.executeUpdate(
+      String.format("DELETE FROM `%s` WHERE %s = ?", this.name, TABLE_COLUMN_KEY),
+      key
+    ) > 0;
+  }
+
+  @Override
+  public @Nullable JsonDocument get(@NonNull String key) {
+    return this.databaseProvider.executeQuery(
+      String.format("SELECT %s FROM `%s` WHERE %s = ?", TABLE_COLUMN_VAL, this.name, TABLE_COLUMN_KEY),
+      resultSet -> resultSet.next() ? JsonDocument.fromJsonString(resultSet.getString(TABLE_COLUMN_VAL)) : null,
+      null,
+      key
+    );
+  }
+
+  @Override
+  public @NonNull List<JsonDocument> find(@NonNull String fieldName, String fieldValue) {
+    return this.databaseProvider.executeQuery(
+      String.format("SELECT %s FROM `%s` WHERE %s LIKE ? ESCAPE '$'", TABLE_COLUMN_VAL, this.name, TABLE_COLUMN_VAL),
+      resultSet -> {
+        List<JsonDocument> jsonDocuments = new ArrayList<>();
+        while (resultSet.next()) {
+          jsonDocuments.add(JsonDocument.fromJsonString(resultSet.getString(TABLE_COLUMN_VAL)));
+        }
+
+        return jsonDocuments;
+      },
+      List.of(),
+      "%\"" + fieldName + "\":" + JsonDocument.GSON.toJson(fieldValue).replaceAll("([_%])", "\\$$1") + "%"
+    );
+  }
+
+  @Override
+  public @NonNull List<JsonDocument> find(@NonNull Map<String, String> filters) {
+    var stringBuilder = new StringBuilder("SELECT ")
+      .append(TABLE_COLUMN_VAL).append(" FROM `")
+      .append(this.name).append('`');
+
+    Collection<String> collection = new ArrayList<>();
+    if (filters.size() > 0) {
+      stringBuilder.append(" WHERE ");
+
+      var iterator = filters.entrySet().iterator();
+
+      while (iterator.hasNext()) {
+        var entry = iterator.next();
+
+        stringBuilder.append(TABLE_COLUMN_VAL).append(" LIKE ? ESCAPE '$'");
+        collection.add("%\"" + entry.getKey() + "\":\"" + entry.getValue().replaceAll("([_%])", "\\$$1") + "\"%");
+
+        if (iterator.hasNext()) {
+          stringBuilder.append(" and ");
+        }
+      }
+    }
+
+    return this.databaseProvider.executeQuery(
+      stringBuilder.append(";").toString(),
+      resultSet -> {
+        List<JsonDocument> jsonDocuments = new ArrayList<>();
+        while (resultSet.next()) {
+          jsonDocuments.add(JsonDocument.fromJsonString(resultSet.getString(TABLE_COLUMN_VAL)));
+        }
+
+        return jsonDocuments;
+      },
+      List.of(),
+      collection.toArray()
+    );
+  }
+
+  @Override
+  public @NonNull Collection<String> keys() {
+    return this.databaseProvider.executeQuery(
+      String.format("SELECT %s FROM `%s`;", TABLE_COLUMN_KEY, this.name),
+      resultSet -> {
+        Collection<String> keys = new ArrayList<>();
+        while (resultSet.next()) {
+          keys.add(resultSet.getString(TABLE_COLUMN_KEY));
+        }
+
+        return keys;
+      }, Set.of());
+  }
+
+  @Override
+  public @NonNull Collection<JsonDocument> documents() {
+    return this.databaseProvider.executeQuery(
+      String.format("SELECT %s FROM `%s`;", TABLE_COLUMN_VAL, this.name),
+      resultSet -> {
+        Collection<JsonDocument> documents = new ArrayList<>();
+        while (resultSet.next()) {
+          documents.add(JsonDocument.fromJsonString(resultSet.getString(TABLE_COLUMN_VAL)));
+        }
+
+        return documents;
+      }, Set.of());
+  }
+
+  @Override
+  public @NonNull Map<String, JsonDocument> entries() {
+    return this.databaseProvider.executeQuery(
+      String.format("SELECT * FROM `%s`;", this.name),
+      resultSet -> {
+        Map<String, JsonDocument> map = new WeakHashMap<>();
+        while (resultSet.next()) {
+          map.put(resultSet.getString(TABLE_COLUMN_KEY),
+            JsonDocument.fromJsonString(resultSet.getString(TABLE_COLUMN_VAL)));
+        }
+
+        return map;
+      }, Map.of());
+  }
+
+  @Override
+  public void iterate(@NonNull BiConsumer<String, JsonDocument> consumer) {
+    this.databaseProvider.executeQuery(
+      String.format("SELECT * FROM `%s`;", this.name),
+      resultSet -> {
+        while (resultSet.next()) {
+          var key = resultSet.getString(TABLE_COLUMN_KEY);
+          var document = JsonDocument.fromJsonString(resultSet.getString(TABLE_COLUMN_VAL));
+          consumer.accept(key, document);
+        }
+
+        return null;
+      }, null);
+  }
+
+  @Override
+  public void clear() {
+    this.databaseProvider.databaseHandler().handleClear(this);
+    this.databaseProvider.executeUpdate(String.format("TRUNCATE TABLE `%s`", this.name));
+  }
+
+  @Override
+  public long documentCount() {
+    return this.databaseProvider.executeQuery("SELECT COUNT(*) FROM `" + this.name + "`;", resultSet -> {
+      if (resultSet.next()) {
+        return resultSet.getLong(1);
+      }
+      return -1L;
+    }, -1L);
   }
 
   @Override
